requirements.txt
matplotlib==3.10.0
pandas==2.2.3
Pillow==11.1.0
torch==2.4.1
torchvision==0.19.1
